requirements.txt
Flask==0.12.2
pysqlite==2.8.3